Updated Facemark API to match newer node utils
