+++ cc/modules/face/Facemark.h
@@ -1,8 +1,7 @@
-#include "Converters.h"
+#include "NativeNodeUtils.h"
 #include "Mat.h"
 #include "Point.h"
 #include "Rect.h"
-#include "Workers.h"
 #include "macros.h"
 #include <iostream>
 #include <opencv2/face.hpp>
@@ -17,39 +16,25 @@ public:
   virtual void load(std::string) = 0;
 
   static void Init(v8::Local<v8::FunctionTemplate>);
-
-  struct AddTrainingSampleWorker;
+  
   static NAN_METHOD(AddTrainingSample);
   static NAN_METHOD(AddTrainingSampleAsync);
-
-  struct LoadModelWorker;
   static NAN_METHOD(LoadModel);
   static NAN_METHOD(LoadModelAsync);
-
-  struct GetDataWorker;
   static NAN_METHOD(GetData);
   static NAN_METHOD(GetDataAsync);
-
-  struct GetFacesWorker;
   static NAN_METHOD(GetFaces);
   static NAN_METHOD(GetFacesAsync);
+  static NAN_METHOD(SetFaceDetector);
+  static NAN_METHOD(Training);
+  static NAN_METHOD(TrainingAsync);
+  static NAN_METHOD(Fit);
+  static NAN_METHOD(FitAsync);
+  static NAN_METHOD(Save);
+  static NAN_METHOD(Load);
 
   static bool detector(cv::InputArray image, cv::OutputArray faces,
                        Nan::Callback *callback);
-
-  struct SetFaceDetectorWorker;
-  static NAN_METHOD(SetFaceDetector);
-
-  struct TrainingWorker;
-  static NAN_METHOD(Training);
-  static NAN_METHOD(TrainingAsync);
-
-  struct FitWorker;
-  static NAN_METHOD(Fit);
-  static NAN_METHOD(FitAsync);
-
-  static NAN_METHOD(Save);
-  static NAN_METHOD(Load);
 };
 
 #endif
