+++ cc/modules/face/FacemarkLBFParams.h
@@ -1,5 +1,6 @@
 #include "macros.h"
-#include "Converters.h"
+#include "TypeConverters.h"
+#include "ArrayConverters.h"
 #include "Rect.h"
 #include <opencv2/face.hpp>
